We were having performance issues with our GraphQL endpoint at my company. I started using a profiler to track down the issues. Most of the issues were with our code, not this library, but I also took a look at what could be improved in this library. 

I went through a lot of iterations with the profiler and ended up improving a lot things:

* Replace calls to `Apply()` extension method with `foreach`. It was bad for perfomance in some places particularly when they were nested and closures were being created
* Similarly with `Map()` and `Fill()`
* Refactor `EnterLeaveListener` to create generic `MatchingNodeVisitor`. Use `is` to check type, rather than reflection
* Change calls to `Any()` to use `Count` property when possible
* Change calls to  `FirstOrDefault()` to use `Find()` on `List<T>`
* Create lists and dictionaries with an initial capacity where possible
* Use pattern matching rather than `is` + cast
* Add a Dictionary to `Directives` class for faster lookups
* Expose `SelectionSet.Selections` as `IList<ISelection>` to allow access to the `Count`
* In `ExpressionFieldResolver`, avoid creating a new `ResolveFieldContext<TSourceType>` and just cast the `Source`
* Avoid creating nested `Task`s in `MiddlewareResolver.Resolve()`
* Simplify code to retrieve `Task<TResult>.Result` when `TResult` is not known
* Improve performance of `NameFieldResolver` by creating and caching delegates

These are executions times for some of the "benchmark" unit tests before and after:

| Test | Before | After |
|------|------|-------|
| Executes_StarWarsBasicQuery_Performant	|	4,134ms|		2,311ms|
| Executes_MultipleProperties_Are_Performant	|6,639ms|		5,260ms|
| Executes_UnionLists_Are_Performant		|	6,208ms|		5,215ms|
| Executes_SimpleLists_Are_Performant		|	890ms|		844ms|