@johnrutherford 2018-08-31
@RehanSaeed, where are you referring to? Other than in unit tests, I don't see any calls to `Task<TResult>.Result` where the task is not being awaited first.