@RehanSaeed 2018-08-31
Not in this repo, the `IValidationRule` needs to return a `Task`, so that the code in the `Authorization` repo doesn't need to call `GetAwaiter.GetResult()` [here](https://github.com/graphql-dotnet/authorization/blob/e9538907e9f6f53a4030f0e3c2444e59e726d7ea/src/GraphQL.Authorization/AuthorizationValidationRule.cs#L73).