@johnrutherford 2018-08-31
@RehanSaeed, ah yes, I noticed that problem too. That was part of the reason I did https://github.com/graphql-dotnet/server/pull/158.